// src/app/models/serverUserObject.model.ts
export class ServerUserObject {

  id: number;
  numBookings: number;
  reviews: string;
  bookings: string;
  saved: string;
  messages: string;
  photoSource: string;
  yearJoined: number;
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  birthday: string;
  phone: string;
  location: string;

  constructor() {}

}